added funny view if input incorrect links
